% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerizer-package.R
\docType{package}
\name{numerizer-package}
\alias{numerizer}
\alias{numerizer-package}
\title{numerizer: Parse Numbers in Natural Language from Strings}
\description{
To convert numbers appearing in Natural Language string into R numeric data types.This is an R wrapper of Python package `numerizer`.
}
\author{
\strong{Maintainer}: AbdulMajedRaja RS \email{amrrs.data@gmail.com}

}
\keyword{internal}
